requirements.txt
# Web framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0

# Settings
pydantic-settings>=2.0,<3.0

# Database
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.14,<2.0

# Task queue
celery[redis]>=5.4,<6.0

# AWS S3
boto3>=1.35,<2.0

# Utilities
python-multipart>=0.0.9
